+++ src/commands/clear.ts
@@ -8,14 +8,14 @@ import path from "path";
 
 import { logger } from "../utils/logger";
 
-export const cleanThumbnails = (dir: string, isRecursiveCall: boolean = false) => {
+export const dropThumbnails = (dir: string, isRecursiveCall: boolean = false) => {
   // Default parameter set to current directory
   const files = fs.readdirSync(dir);
   files.forEach(file => {
     const filePath = path.join(dir, file);
     const stat = fs.statSync(filePath);
     if (stat.isDirectory()) {
-      cleanThumbnails(filePath, true);
+      dropThumbnails(filePath, true);
     } else {
       if (file.endsWith("_thn.jpg")) {
         fs.unlinkSync(filePath);
@@ -26,7 +26,7 @@ export const cleanThumbnails = (dir: string, isRecursiveCall: boolean = false) =
   if (!isRecursiveCall) logger.info("Asiair dump cleaning - Thumbnails deleted ✅");
 };
 
-export const removeEmptyDirectories = (dir: string) => {
+export const dropEmptyDirectories = (dir: string) => {
   const isDirEmpty = (directory: string): boolean => {
     const files = fs.readdirSync(directory);
     if (files.length === 0) {
