+++ src/commands/preprocess.generate-scripts.ts
